using shared utils file; built level 2 from rom data
