+++ CadEditor/settings_nes/young_indiana_jones_chronicles/Settings_Level1_2_part_1.cs
@@ -1,14 +1,29 @@
 using CadEditor;
 using System;
-using System.Drawing;
+//css_include shared_settings/SharedUtils.cs;
+//css_include young_indiana_jones_chronicles/IndyUtils.cs;
 
 public class Data 
 { 
   public OffsetRec getScreensOffset()  { return new OffsetRec( 0xd3b4, 1 , 12*128, 12, 128);   }
   public bool getScreenVertical()      { return true; }
-  public string getBlocksFilename()    { return "young_indiana_jones_chronicles_1_2.png"; }
+  public bool isBuildScreenFromSmallBlocks() { return true; }
   
   public bool isBigBlockEditorEnabled() { return false; }
-  public bool isBlockEditorEnabled()    { return false; }
+  public bool isBlockEditorEnabled()    { return true; }
   public bool isEnemyEditorEnabled()    { return false; }
+  
+  public OffsetRec getBlocksOffset()    { return new OffsetRec(0xCDB4 , 1  , 0x1000);  }
+  public int getPalBytesAddr()          { return 0xd9b4; }
+  public GetBlocksFunc        getBlocksFunc() { return IndyUtils.getBlocksFromTiles16Pal1V;}
+  public SetBlocksFunc        setBlocksFunc() { return IndyUtils.setBlocksFromTiles16Pal1V;}
+  public int getBigBlocksCount() { return 96; }
+  public int getBlocksCount()    { return 96; }
+  
+  public GetVideoPageAddrFunc getVideoPageAddrFunc() { return SharedUtils.fakeVideoAddr(); }
+  public GetVideoChunkFunc    getVideoChunkFunc()    { return SharedUtils.getVideoChunk("chr2.bin");    }
+  public SetVideoChunkFunc    setVideoChunkFunc()    { return null; }
+  
+  public GetPalFunc           getPalFunc()           { return SharedUtils.readPalFromBin("pal2.bin"); }
+  public SetPalFunc           setPalFunc()           { return null;}
 }
